Added better function querying + functionality
Querying is now more dynamic based on the function that has the @FunctionalMod annotation. The querying now separates the name, parameters, and return type. That made it possible to inject or provide values for function parameters. Currently only supports BoxlinContext. When returning anything other than void (Unit) the executor picks it up and uses that as the mod "instance" otherwise it still goes back to VirtualModInstance.